Удален файл MovieQuizUITests/MovieQuizUITestsLaunchTests.swift Сделан отдельный файл struct StubNetworkClient с данными для тестирования Изменен уровень доступности func yesButtonClicked и func noButtonClicked на internal, т.к. используют внутри не приватные методы
